Main 4, 5, 6

+++ My first Java Project/src/com/company/main3.java
@@ -5,11 +5,11 @@ import java.util.Scanner;
 public class main3 {
     public static void main(String[] args){
         Scanner scn = new Scanner (System.in);
-        int num, num2, num3, result;
+        float num, num2, num3, result;
         System.out.println("Please enter first number:");
-        num = scn.nextInt();
+        num = scn.nextFloat();
         System.out.println("Please enter second number:");
-        num2 = scn.nextInt();
+        num2 = scn.nextFloat();
         result = num + num2;
         System.out.print("Sum of two number is: " + result);
     }
